Redirect clib functions that use FILE*
StarCraft has its own statically linked clib, which coexists with the more
modern implementation of Magnetar. Most of the time this is not a problem,
however when state is passed between the two clib versions, things break.
One example is functions that use FILE*. By redirecting all calls to a
single version, the problem is fixed.